# Conflicts:
#	pom.xml
#	src/main/java/app/controllers/AuctionController.java
#	src/main/java/app/controllers/BidController.java
#	src/main/java/app/controllers/FirstScene.java
#	src/main/java/app/controllers/LiveController.java
#	src/main/java/app/controllers/MessController.java
#	src/main/java/app/controllers/MineController.java
#	src/main/java/app/controllers/MyHistoryController.java
#	src/main/java/app/dao/BidDAO.java
#	src/main/java/app/dao/HistoryDAO.java
#	src/main/java/app/dao/ItemDAO.java
#	src/main/java/app/models/Auction.java
#	src/main/java/app/network/Client.java
#	src/main/java/app/network/PlaceBidCommand.java
#	src/main/java/app/service/AuctionService.java
#	src/main/java/app/service/ItemService.java
#	src/main/resources/app/views/ConnectServerController.fxml
#	src/main/resources/app/views/firstscene.fxml
#	src/main/resources/app/views/live_auction.fxml
#	src/main/resources/app/views/login_scene.fxml
#	src/main/resources/app/views/mess_chat.fxml
#	src/main/resources/app/views/my_auction.fxml
#	src/main/resources/app/views/register_account.fxml
#	src/main/resources/schema.sql

+++ src/main/java/app/models/Bidder.java
@@ -1,6 +1,7 @@
 package app.models;
 
 import app.enums.UserRole;
+import app.observer.AuctionObserver;
 
 public class Bidder extends User implements AuctionObserver {
   public Bidder(int id, String name, Account account, Wallet wallet) {
@@ -17,7 +18,7 @@ public class Bidder extends User implements AuctionObserver {
   }
 
   @Override
-  public void onNewBidPlaced(String itemName, double newPrice, String bidderName) {
+  public void onNewBidPlaced(String itemName, long newPrice, String bidderName) {
     if (this.getName().equals(bidderName)) {
       System.out.println("[" + getName() + "]: Bạn đã giữ giá cao nhất mới: $" + newPrice);
     } else {
@@ -34,7 +35,7 @@ public class Bidder extends User implements AuctionObserver {
   }
 
   @Override
-  public void onAuctionClosed(String itemName, String winnerName, double finalPrice) {
+  public void onAuctionClosed(String itemName, String winnerName, long finalPrice) {
     if (this.getName().equals(winnerName)) {
       System.out.println(
           "["
